Rewrite pgm_read_* macros, fixing bugs and speeding up
The pgm_read_xxxx_unaligned macros contained bugs which allowed GCC to
optimize things in such a way it could still perform unaligned accesses
to PMEM, causing random crashes.

Rewrite the pgm_read_dword_unaligned in assembly and using the funnel
shifter (as per the ISA reference 3.3.2, pg25/26, adjusting for the fact
we do have an unaligned access exception) so that it runs as fast as the
original pgm_read_byte macro (6 instructions).

Also, optimize pgm_read_byte in the same way, speeding it up by 17% by
removing an instruction.