Move to a14 for temp in dword-unaligned, __pstr__ tag
Fixes #3

Change from a15 to a14 in the pgm_read_dword_unaligned macro to allow
-O0 builds to complete.

Change the local __c to __pstr__ to make it easier/safer to match in
link stage in the case when it is not automatically included in PROGMEM
(i.e. in templates where section attrs get tossed).